backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-login==0.6.2
flask-sqlalchemy==3.1.1
flask-admin==1.6.1
flask-wtf==1.2.1
email_validator==2.2.0
werkzeug==2.3.7
pyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
openai==0.28.0
selenium==4.11.2
python-dotenv==1.0.0
anthropic==0.8.0
Gunicorn==21.2.0